move(cosmo): move utils to private API
