+++ astropy/cosmology/flrw/w0cdm.py
@@ -5,7 +5,7 @@
 import numpy as np
 from numpy import sqrt
 
-from astropy.cosmology._utils import aszarr, deprecated_keywords
+from astropy.cosmology._src.utils import aszarr, deprecated_keywords
 from astropy.cosmology.core import dataclass_decorator
 from astropy.cosmology.parameter import Parameter
